fix: code clean up
icons need clean up, but code is nicely broken up now.
