+++ src/modules/account/AccountCarousel.vue
@@ -6,91 +6,21 @@
 <script setup lang="ts">
 import "vue3-carousel/dist/carousel.css";
 import { Carousel, Slide } from "vue3-carousel";
-import AccountIcon from "@/components/icons/solid/AccountIcon.vue";
-import UnlinkIcon from "@/components/icons/outline/UnlinkIcon.vue";
+import type { Account } from "@/modules/account/Account";
+import AccountItem from "@/modules/account/AccountItem.vue";
 
-// defineProps({
-//   accounts: Array<{
-//     type: String;
-//     username: String;
-//   }>,
-// });
-
-const accounts = [
-  {
-    type: AccountIcon,
-    username: "mike@mytiki.com",
+defineProps({
+  accounts: {
+    type: Array<Account>,
+    required: true,
   },
-  {
-    type: AccountIcon,
-    username: "mike@mytiki.com",
-  },
-  {
-    type: AccountIcon,
-    username: "mike@mytiki.com",
-  },
-  {
-    type: AccountIcon,
-    username: "mike@mytiki.com",
-  },
-];
+});
 </script>
 
 <template>
   <carousel :items-to-show="3.5" snapAlign="start">
     <slide v-for="account in accounts" :key="accounts">
-      <div>
-        <div class="account-button" role="button">
-          <component :is="account.type" class="icon" />
-          <div class="unlink"><unlink-icon class="unlink-icon" /> Unlink</div>
-        </div>
-        <p class="username">{{ account.username }}</p>
-      </div>
+      <account-item :account="account" />
     </slide>
   </carousel>
 </template>
-
-<style scoped>
-.account-button {
-  background-color: var(--tiki-primary-background-color);
-  font-size: var(--tiki-font-size);
-  border-radius: 0.625em;
-  box-shadow: 0 4px 4px 0 rgba(0, 0, 0, 5%);
-  display: block;
-}
-
-.icon {
-  height: 2.75em;
-  margin: 1em 0 0.571em 0;
-}
-
-.unlink {
-  font-family: var(--tiki-font-family);
-  font-size: var(--tiki-font-size-sm);
-  line-height: var(--tiki-line-height-sm);
-  font-weight: bold;
-  color: var(--tiki-secondary-text-color);
-  display: flex;
-  align-items: center;
-  padding: 0 0.571em 0.571em 0.571em;
-}
-
-.unlink-icon {
-  margin-right: 0.571em;
-  height: 1.714em;
-}
-
-.username {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  font-family: var(--tiki-font-family);
-  font-size: var(--tiki-font-size-xs);
-  line-height: var(--tiki-line-height-xs);
-  font-weight: 500;
-  color: var(--tiki-secondary-text-color);
-  width: 7.5em;
-  display: block;
-  margin: 0.5em 0;
-}
-</style>
