+++ src/modules/account/AccountSheet.vue
@@ -4,36 +4,49 @@
   -->
 
 <script setup lang="ts">
-import CloseIcon from "@/components/icons/outline/CloseIcon.vue";
-import CircleButton from "@/components/buttons/CircleButton.vue";
-import BackButton from "@/components/buttons/BackButton.vue";
 import AccountForm from "@/modules/account/AccountForm.vue";
 import TextButton from "@/components/buttons/TextButton.vue";
 import AccountIcon from "@/components/icons/outline/AccountIcon.vue";
 import AccountCarousel from "@/modules/account/AccountCarousel.vue";
+import { Account } from "@/modules/account/Account";
+import { AccountType } from "@/modules/account/AccountType";
+import HeaderBack from "@/components/HeaderBack.vue";
+import CloseIcon from "@/components/icons/outline/CloseIcon.vue";
+import CircleButton from "@/components/buttons/CircleButton.vue";
 
 defineEmits(["close", "back"]);
+
+const accounts: Array<Account> = [
+  {
+    type: AccountType.GMAIL,
+    username: "mike@mytiki.com",
+  },
+  {
+    type: AccountType.GMAIL,
+    username: "mike@mytiki.com",
+  },
+  {
+    type: AccountType.GMAIL,
+    username: "mike@mytiki.com",
+  },
+  {
+    type: AccountType.GMAIL,
+    username: "mike@mytiki.com",
+  },
+];
 </script>
 
 <template>
-  <div class="heading">
-    <back-button text="Rewards" @click="$emit('back')" />
+  <header-back text="Rewards" @back="$emit('back')">
     <circle-button @click="$emit('close')" :icon="CloseIcon" />
-  </div>
+  </header-back>
   <account-form />
   <p class="linked-accounts">Linked Accounts</p>
-  <account-carousel class="account-carousel" />
+  <account-carousel :accounts="accounts" class="account-carousel" />
   <text-button text="Link Account" :icon="AccountIcon" />
 </template>
 
 <style scoped>
-.heading {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  margin: 0 0.5em 2em 0.5em;
-}
-
 .linked-accounts {
   font-family: var(--tiki-font-family);
   font-size: var(--tiki-font-size);
