+++ src/modules/account/AccountType.ts
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+import AccountIcon from "@/components/icons/outline/AccountIcon.vue";
+
+export enum AccountType {
+  GMAIL = "gmail",
+}
+
+export const icon = (type: AccountType): Object => {
+  switch (type) {
+    case AccountType.GMAIL:
+      return AccountIcon;
+    default:
+      return AccountIcon;
+  }
+};
