+++ src/modules/history/HistorySheet.vue
@@ -1,0 +1,20 @@
+<!--
+  - Copyright (c) TIKI Inc.
+  - MIT license. See LICENSE file in root directory.
+  -->
+
+<script setup lang="ts">
+import HistoryDetails from "@/modules/history/HistoryDetails.vue";
+import HeaderBack from "@/components/HeaderBack.vue";
+import CloseIcon from "@/components/icons/outline/CloseIcon.vue";
+import CircleButton from "@/components/buttons/CircleButton.vue";
+
+defineEmits(["close", "back"]);
+</script>
+
+<template>
+  <header-back text="Rewards" @back="$emit('back')">
+    <circle-button @click="$emit('close')" :icon="CloseIcon" />
+  </header-back>
+  <history-details />
+</template>
