+++ src/modules/offer/OfferSheet.vue
@@ -6,12 +6,12 @@
 <script setup lang="ts">
 import { Offer } from "./Offer";
 import type { PropType } from "vue";
-import SheetTitle from "@/components/SheetTitle.vue";
 import QuestionIcon from "@/components/icons/outline/QuestionIcon.vue";
 import TextButton from "@/components/buttons/TextButton.vue";
-import OfferCard from "@/modules/offer/components/OfferCard.vue";
 import CircleButton from "@/components/buttons/CircleButton.vue";
-import OfferDataUse from "@/modules/offer/components/OfferDataUse.vue";
+import OfferDataUse from "@/modules/offer/OfferDataUse.vue";
+import Card from "@/components/Card.vue";
+import HeaderTitle from "@/components/HeaderTitle.vue";
 
 defineEmits(["close", "learn", "accept"]);
 defineProps({
@@ -23,11 +23,10 @@ defineProps({
 </script>
 
 <template>
-  <div class="heading">
-    <sheet-title title="Data Rewards" subtitle="Share data. Earn rewards." />
+  <header-title title="Data Rewards" subtitle="Share data. Earn rewards.">
     <circle-button :icon="QuestionIcon" @click="$emit('learn')" />
-  </div>
-  <offer-card :src="offer.reward" :description="offer.description" />
+  </header-title>
+  <card :image="offer.image" :description="offer.description" />
   <offer-data-use :bullets="offer.bullets" class="data-use" />
   <div class="footer">
     <text-button
@@ -41,13 +40,6 @@ defineProps({
 </template>
 
 <style scoped>
-.heading {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  margin: 0 0.5em 2em 0.5em;
-}
-
 .footer {
   display: flex;
   align-items: center;
