+++ src/modules/reward/Reward.ts
@@ -3,10 +3,7 @@
  * MIT license. See LICENSE file in root directory.
  */
 
-import type { Bullet } from "@/modules/offer/Bullet";
-
-export interface Offer {
+export interface Reward {
   image: String;
   description: String;
-  bullets: Array<Bullet>;
 }
