--- src/modules/reward/Reward.vue
@@ -1,107 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<script setup lang="ts">
-import SheetTitle from "@/components/SheetTitle.vue";
-import CloseIcon from "@/components/icons/outline/CloseIcon.vue";
-import Counter from "@/modules/reward/Counter.vue";
-import ConfettiExplosion from "vue-confetti-explosion";
-import CircleButton from "@/components/buttons/CircleButton.vue";
-import SquareButton from "@/components/buttons/SquareButton.vue";
-import AccountIcon from "@/components/icons/solid/AccountIcon.vue";
-import RedeemIcon from "@/components/icons/outline/RedeemIcon.vue";
-import TextButton from "@/components/buttons/TextButton.vue";
-import ScanIcon from "@/components/icons/outline/ScanIcon.vue";
-import HistorySummary from "@/modules/history/HistorySummary.vue";
-import ReceiptIcon from "@/components/icons/solid/ReceiptIcon.vue";
-import OfferCarousel from "@/modules/offer/components/OfferCarousel.vue";
-
-defineEmits(["close", "history", "account"]);
-</script>
-
-<template>
-  <div class="heading">
-    <sheet-title title="Data Rewards" subtitle="Share data. Earn rewards." />
-    <circle-button @click="$emit('close')" :icon="CloseIcon" />
-  </div>
-  <ConfettiExplosion class="confetti" />
-  <counter :amount="2750" class="counter" />
-  <history-summary
-    class="history-summary"
-    :amount="25"
-    :date="new Date()"
-    :icon="ReceiptIcon"
-    @click="$emit('history')"
-  />
-  <div>
-    <p class="offers">Offers</p>
-    <offer-carousel
-      :offers="[
-        {
-          src: './src/assets/offer-card.png',
-          description:
-            'Share your receipts to earn rewards like gift cards or cash!',
-        },
-        {
-          src: './src/assets/offer-card.png',
-          description:
-            'Share your receipts to earn rewards like gift cards or cash!',
-        },
-        {
-          src: './src/assets/offer-card.png',
-          description:
-            'Share your receipts to earn rewards like gift cards or cash!',
-        },
-      ]"
-    />
-  </div>
-  <div class="button-row">
-    <square-button
-      :icon="AccountIcon"
-      text="Link Accounts"
-      class="account-icon"
-      @click="$emit('account')"
-    />
-    <square-button :icon="RedeemIcon" text="Redeem Points" />
-  </div>
-  <text-button text="Scan Receipt" :icon="ScanIcon" />
-</template>
-
-<style scoped>
-.heading {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  margin: 0 0.5em 2em 0.5em;
-}
-
-.button-row {
-  display: flex;
-  align-items: start;
-  margin: 2em auto;
-  width: fit-content;
-}
-
-.account-icon {
-  margin-right: 3.125em;
-}
-
-.confetti {
-  margin: auto;
-}
-
-.history-summary {
-  margin: 2em 0;
-}
-
-.offers {
-  font-family: var(--tiki-font-family);
-  font-size: var(--tiki-font-size);
-  line-height: var(--tiki-line-height);
-  color: var(--tiki-primary-text-color);
-  text-transform: uppercase;
-  font-weight: bold;
-}
-</style>
